backend/requirements.txt
# Growin App - Backend Requirements
# Updated: 2026-01-18
# For Python >= 3.10

## Core Web Framework
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0
pydantic>=2.5.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.26.0,<0.28.0

## LangChain & LLM Support
langchain>=0.1.0,<0.2.0
langchain-core>=0.1.0,<0.2.0
langchain-openai>=0.0.5,<0.1.0
langchain-anthropic>=0.1.0,<0.2.0
langchain-google-genai>=0.0.11,<0.1.0
langchain-ollama>=0.0.1,<0.1.0
langgraph>=0.0.20,<0.1.0

## Trading & Market Data APIs
alpaca-py>=0.15.0,<0.16.0
yfinance>=0.2.35,<0.3.0
finnhub-python>=2.4.19,<3.0.0

## Data Processing & Analysis
pandas>=2.1.4,<3.0.0
numpy>=1.26.3,<2.0.0
scikit-learn>=1.3.0,<2.0.0

## Technical Analysis
# TA-Lib requires system installation:
# macOS: brew install ta-lib && pip install TA-Lib
# Linux: sudo apt-get install ta-lib && pip install TA-Lib
# Then: pip install TA-Lib

## ML/AI & Transformers
torch>=2.1.0,<3.0.0
transformers>=4.36.0,<5.0.0

## Apple Silicon (MLX) - SOTA for M-series Macs
mlx>=0.0.9,<0.1.0
mlx-lm>=0.0.9,<0.1.0

## Analytics Database - OLAP
duckdb>=0.9.0,<0.10.0

## Model Context Protocol (MCP)
mcp>=0.1.0,<1.0.0

## Hugging Face
huggingface-hub>=0.20.0,<1.0.0

## Search & Fuzzy Matching
rapidfuzz>=3.0.0,<4.0.0

## Development & Testing
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.21.1,<0.22.0
pylint>=3.0.3,<4.0.0

## News & Sentiment Analysis
newsapi-python>=0.2.7,<0.3.0
tavily-python>=0.3.0,<0.4.0
vaderSentiment>=3.3.2,<4.0.0

## Time Series Forecasting (TTM-R2)
# Note: Requires Python 3.9-3.12
granite-tsfm @ git+https://github.com/ibm-granite/granite-tsfm.git@v0.2.22
psutil>=5.9.0
